Added an FAB for create game
When the user uses mobile, the 'create game' option is removed from the side menu/bottom menu and added as a FAB to the game list.

This saves main navigation space (and confusion) for an option that is not often used. We should actually try to NOT let people first make games, but join one first.

On desktop we have more than enough real estate though, so can keep it in the side menu.